pyproject.toml
[project]
name = "ai4ts"
description = "Utilities for analyzing AI time series models"
version = "2024.09.001"
authors = [
    {name = "Bryce Allen", email = "ballen@anl.gov"},
]
keywords = ["ai", "time series"]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
  "fastparquet",
  "huggingface_hub",
  "matplotlib",
  "numpy",
  "pandas",
  "pyarrow",
  "pyyaml",
  "statsforecast",
  "statsmodels",
  "ujson",
]

[project.optional-dependencies]
chronos = [
  "chronos @ git+https://github.com/amazon-science/chronos-forecasting.git",
]
lag-llama = [
  "lag-llama @ git+https://github.com/time-series-foundation-models/lag-llama.git@main",
]
timesfm = ["timesfm"]
lint = ["black"]
ml = ["ai4ts[chronos,lag-llama,timesfm]"]
all = ["ai4ts[chronos,lag-llama,timesfm,lint]"]

[project.urls]
Homepage = "https://github.com/bd4/ai-science-time-series"

[project.scripts]
ai4ts-arma-gen = "ai4ts.arma:main"
ai4ts-compare = "ai4ts.compare:main"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
per-file-ignores = { "__init__.py" = ['F401'] }
